### 2025-11-03 #9051 @rkurchin
Hi @pfaendtner, I'm one of the AEiC's for the BCM track and your paper has been transferred here as it's a better fit. However, we have a bit of a backlog at the moment and so I'll have to waitlist this submission for now until an appropriate editor becomes available.

In the meantime, I can see you've already addressed a few of the initial issues flagged in our automatic checks, thanks, that will help make for a smooth review process once things get underway! One other thing in that category is that the paper is a bit on the long side; JOSS typically targets 1000 words or less. Often, this means that some paper content would be a better fit in the package documentation, so maybe you can consider that as an option?
