### 2025-11-10 #9051 @espottesmith
@mitwood @jkalayan @jeevster can any of y'all offer your time and expertise to review this paper/codebase?
