### 2025-11-17 #9051 @espottesmith
@PythonFZ @CompRhys could either of y'all contribute reviews to this manuscript/codebase?
