### 2025-11-19 #9468 @editorialbot
[REVIEW]: SPARC: An Automated Workflow Toolkit for Accelerated Active Learning of Reactive Machine Learning Interatomic Potentials
**Submitting author:** @pfaendtner (<a href="http://orcid.org/0000-0001-6727-2957">Jim Pfaendtner</a>)
**Repository:** https://github.com/rahulumrao/sparc
**Branch with paper.md** (empty if default branch): main
**Version:** 0.1
