### 2026-01-26 #9468 @mitwood
Revisiting this as I realized some of the check boxes in the review template where wrongfully left unchecked, my apologies. The only things left unchecked are where I am requesting changes to be made.
1) Contribution - Please revisit your splash page for the repo and append a section with contributing authors/developers
2) Functionality - This is coupled with the next point, but I do not feel that the structure of the code is discussed in the documentation or accompanying paper. What are the main functions, and the 'flow' of the codes' execution?
3) Quality of writing - My concern is for developers or users that want to change parts of the code to suit their needs, or contribute back their changes to this project. There isn't an explanation of the main objects (parallel execution or not) in the code, and what should/shouldn't be modified. There are core parts of any code that extra care should be taken before making changes, and other sections that can be appended without any disruption to the main workflow. Knowing how SPARC is structured (giving labels to code execution paths) will be beneficial for users and developers to know.
