+++ paper/ref.bib
@@ -172,4 +172,30 @@
   doi={10.1002/jcc.21057}
 }
 
+@article{david2025arcann,
+  author  = {David, Rolf and de la Puente, Miguel and Gomez, Axel and Anton, Olaia and Stirnemann, Guillaume and Laage, Damien},
+  title   = {{ArcaNN}: automated enhanced sampling generation of training sets for chemically reactive machine learning interatomic potentials},
+  journal = {Digital Discovery},
+  year    = {2025},
+  volume  = {4},
+  pages   = {54--72},
+  doi     = {10.1039/D4DD00209A}
+}
 
+@software{kasoar2024januscore,
+  author  = {Kasoar, Elliott and Zanca, Federica and Devereux, Harvey and Austin, Patrick and Mason, David and Wilkins, Jacob and Elena, Alin Marin},
+  title   = {{janus-core}: Tools for machine learnt interatomic potentials},
+  year    = {2024},
+  doi     = {10.5281/zenodo.13152495},
+  url     = {https://github.com/stfc/janus-core}
+}
+
+@article{vandermause2020flare,
+  author  = {Vandermause, Jonathan and Torrisi, Steven B. and Batzner, Simon and Xie, Yu and Sun, Lixin and Kolpak, Alexie M. and Kozinsky, Boris},
+  title   = {On-the-fly active learning of interpretable {Bayesian} force fields for atomistic rare events},
+  journal = {npj Computational Materials},
+  year    = {2020},
+  volume  = {6},
+  pages   = {20},
+  doi     = {10.1038/s41524-020-0283-z}
+}
