+++ AUTHORS.md
@@ -1,5 +1,0 @@
-## SPARC Contributors
-
-- Rahul Verma
-- Nisarg Joshi
-- Jim Pfaendtner
