+++ AUTHORS.md
@@ -1,0 +1,5 @@
+## SPARC Contributors
+
+- Rahul Verma
+- Nisarg Joshi
+- Jim Pfaendtner
