### 2026-03-17 #9468 @PythonFZ
I have finished my review at https://github.com/rahulumrao/sparc/issues/4.

>  Has the submitting author (pfaendtner) made major contributions to the software?

There are no commits but I assume JOSS has some guidelines for project supervision.
All other points have been resolved.
