### 2026-03-17 #9468 @espottesmith
If a supervisor has made a significant intellectual contribution (e.g., guiding design decisions), then they can be listed as an author. From the JOSS guidelines:

> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are.
