+++ examples/README.md
@@ -49,19 +49,19 @@ wget https://zenodo.org/records/18261342/files/AmmoniaBorate.tar.gz?download=1
 tar -xvf AmmoniaBorate.tar.gz
 ```
 
-<!-- ## Citation
+## Citation
 
-If you use this dataset or workflow in your research, please cite:
+If you use this dataset in your research, please cite:
 
 ```bibtex
 @software{sparc,
   author = {Verma, Rahul and Joshi, Nisarg and Pfaendtner, Jim},
-  doi    = {10.xxxx/yyyy},
+  doi    = {https://doi.org/10.5281/zenodo.18261342},
   license = {MIT},
   month  = {jan},
   title  = {{SPARC}: An Automated Workflow Toolkit for Accelerated Active Learning of Reactive Machine Learning Interatomic Potentials},
-  url    = {https://github.com/rahulumrao/sparc},
+  url    = {https://zenodo.org/records/18261342},
   year   = {2025}
 }
-``` -->
+```
 
