### 2026-03-31 #9468 @rahulumrao
Hi @espottesmith,

Thanks for sending the checklist. We have reviewed the authors checklist please see the details below:

- The author and their affiliations (including ORCIDs) are correct.
- The version number for our final code is version 0.1.0
- We have posted the Zenodo DOI on our Github (sparc/examples/README.md)
- The title, author list and license remain the same in the software as it appears in the archive.

Please let us know if we can provide any other information on our end.
