### 2026-04-01 #9468 @espottesmith
Thanks, @rahulumrao.

Your Zenodo archive needs to have a title and author list identical to your paper. If the DOI in your README (https://doi.org/10.5281/zenodo.18261342) is correct, this is not the case. The license also appears to be different between your code (MIT) and the archive (CC-BY).
