### 2026-04-01 #9468 @espottesmith
Comments on your paper:

- "The workflow16 consists of three main steps, which are executed iteratively in a loop until a reactive and stable17
MLIP is constructed." <-- wording is kind of confusing. Surely they're all reactive potentials? Maybe it should be "until a stable reactive MLIP is constructed" or even just "until a stable MLIP"?
- It should be clear that  "(iter_000000, iter_000001,)" are just examples. You could add an "e.g.," before, or an ellipsis ("...") after
