+++ paper/paper.md
@@ -31,7 +31,7 @@ bibliography: ref.bib
 
 Machine learning interatomic potentials (MLIPs) are rapidly becoming an essential modeling tool, as they allow atomistic simulations to achieve larger systems and longer time scales, while retaining the accuracy of ab-initio methods. However, constructing a reactive and transferable MLIP is still challenging, as it requires high quality training data including rare events, high energy intermediates and transition states.
 
-Herein, we present `SPARC` (Smart Potential with Atomistic Rare events and Continuous learning), a modular Python workflow which automates the construction of reactive MLIPs that can generate chemically accurate potential energy surfaces (PES) for rare events.  The idea for `SPARC` is to use an active learning (AL) protocol coupled with advanced sampling techniques that systematically identify new configurations and train ML models on-the-fly. The workflow consists of three main steps, which are executed iteratively in a loop until a reactive and stable MLIP is constructed. `SPARC` also includes utilities for visualizing configurations and plotting various properties to monitor the workflow across iterations.
+Herein, we present `SPARC` (Smart Potential with Atomistic Rare events and Continuous learning), a modular Python workflow which automates the construction of reactive MLIPs that can generate chemically accurate potential energy surfaces (PES) for rare events.  The idea for `SPARC` is to use an active learning (AL) protocol coupled with advanced sampling techniques that systematically identify new configurations and train ML models on-the-fly. The workflow consists of three main steps, which are executed iteratively in a loop until a stable reactive MLIP is constructed. `SPARC` also includes utilities for visualizing configurations and plotting various properties to monitor the workflow across iterations.
 
 # Statement of need
 
@@ -48,7 +48,7 @@ Several groups have coupled AL with enhanced sampling or pathway exploration tec
 # Features and Implementation
 
 
-`SPARC` is written in Python and can run on both local workstations and high-performance clusters. We utilized the Atomic Simulation Environment (ASE) [@AtomicSimulationEnvironment] library as the brain for `SPARC` which enables seamless integration of the MD engine, machine learning architecture, and electronic structure codes. To ensure portability, `SPARC` implements a file-based management system, and the output of each stage is written in a structured directory. The workflow iterations are stored in subdirectories (`iter_000000`, `iter_000001`,), with subfolders for reference first principle (FP) calculations (`00.dft`), MLIP training (`01.train`), and ML/MD simulations (`02.dpmd`). The `SPARC` workflow is controlled through a single YAML configuration file.
+`SPARC` is written in Python and can run on both local workstations and high-performance clusters. We utilized the Atomic Simulation Environment (ASE) [@AtomicSimulationEnvironment] library as the brain for `SPARC` which enables seamless integration of the MD engine, machine learning architecture, and electronic structure codes. To ensure portability, `SPARC` implements a file-based management system, and the output of each stage is written in a structured directory. The workflow iterations are stored in subdirectories (e.g. `iter_000000`, `iter_000001`,), with subfolders for reference first principle (FP) calculations (`00.dft`), MLIP training (`01.train`), and ML/MD simulations (`02.dpmd`). The `SPARC` workflow is controlled through a single YAML configuration file.
 
 `SPARC` currently supports VASP [@hafner2008vasp] and CP2K [@hutter2014cp2k] for electronic structure calculation, although this could be readily expanded in the future. For MLIP training we use DeePMD-kit architecture [@wang2018deepmd], with the ensemble model approach for query-by-committee (QbC) uncertainty estimation [@miksch2021strategiesMLPs]. ML/MD simulations are run using ASE MD engine coupled together with both DeePMD and PLUMED calculator.
 
