### 2026-04-01 #9468 @rahulumrao
Thank you @espottesmith for pointing it out.
I have fixed the title and author list in the Zenodo record and changed the license to MIT to keep it consistent.
I have also implemented the suggested changes in the manuscript draft.

Please let us know if something else is required from our side.
