### 2026-04-02 #9468 @espottesmith
Thanks for making the changes, @rahulumrao.

The affiliations still aren't exactly the same (your paper includes department names, while your Zenodo archive does not). Also, the version isn't the same (Zenodo is v1; you told me the version that should be tied to your manuscript is v0.1.0

Oh, and I just noticed that your archive doesn't actually contain a copy of your code. It seems to only contain input files. The point of the archive is that, whatever happens to your GitHub repo, there will always be a record of your code that is directly tied to your JOSS manuscript. My apologies. I should have mentioned this first thing, but I wasn't looking closely at the archive contents.
