+++ README.md
@@ -279,18 +279,28 @@ pre-commit run --all-files
 > Also, this code is designed to work in a Linux environment. It may not be fully compatible with macOS systems.
 
 
-<!-- ## Citation
+## Citation
 
-If you use this dataset or workflow in your research, please cite:
+If you use this software or the dataset in your research, please cite:
 
 ```bibtex
 @software{sparc,
   author = {Verma, Rahul and Joshi, Nisarg and Pfaendtner, Jim},
-  doi    = {10.xxxx/yyyy},
+  doi    = {https://doi.org/10.5281/zenodo.19389278},
   license = {MIT},
   month  = {jan},
   title  = {{SPARC}: An Automated Workflow Toolkit for Accelerated Active Learning of Reactive Machine Learning Interatomic Potentials},
   url    = {https://github.com/rahulumrao/sparc},
   year   = {2025}
 }
-``` -->
+
+@dataset{sparc,
+  author = {Verma, Rahul and Joshi, Nisarg and Pfaendtner, Jim},
+  doi    = {https://doi.org/10.5281/zenodo.18261342},
+  license = {MIT},
+  month  = {jan},
+  title  = {{SPARC}: An Automated Workflow Toolkit for Accelerated Active Learning of Reactive Machine Learning Interatomic Potentials},
+  url    = {https://zenodo.org/records/18261342},
+  year   = {2025}
+}
+```
