### 2026-04-02 #9468 @rahulumrao
Hi @espottesmith. Thank you for the feedback.

I have linked the Zenodo record with GitHub code release. Also updated the README file with citation and DOI.

Please let me know if anything else needs to be adjusted.
