### 2026-04-13 #9468 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
