### 2026-04-17 #9468 @rahulumrao
Hi @espottesmith

I wanted to reach out to know the status of the paper. Should we consider the paper accepted for publication or are there any other steps on your end that need to be taken care of?

We appreciate your time and consideration.
