### 2026-04-19 #9468 @espottesmith
Hey @rahulumrao,

The paper isn't accepted until it's accepted. We need to wait for an Associate Editor in Chief to take a look and give final approval. Like the rest of the review, that process will take place in this Issue thread. But there's nothing more that you need to do at this point.
