+++ scripts/ComputeBias.py
@@ -1,0 +1,249 @@
+"""
+ComputeBias.py — Recover physical forces from biased MD simulations.
+
+This module contains the core physics for removing bias forces introduced
+by enhanced sampling methods (metadynamics, umbrella sampling, etc.)
+acting through collective variables (CVs).
+
+
+The problem
+-----------
+In a biased simulation the force on each atom is not the true physical
+force. The MD engine (e.g. CP2K) reports:
+
+    F_total(t) = F_physical(t) + F_bias(t)
+
+We want F_physical for training machine-learning interatomic potentials.
+
+
+How the bias acts
+-----------------
+The bias potential V_bias depends on atomic positions only through
+collective variables s_k(R):
+
+    V_bias = V_bias(s_1, s_2, ..., s_K)
+
+The bias force on atom i from all K collective variables follows from
+the chain rule:
+
+                    K
+    F_bias_i(t) = SUM   f_k(t)  *  (ds_k/dR_i)(t)
+                   k=1
+
+where:
+
+    f_k(t) = -dV_bias/ds_k      scalar force on CV k (from DUMPFORCES)
+    ds_k/dR_i                    how CV k changes when atom i moves
+                                 (from DUMPDERIVATIVES)
+
+This is general — it works for any CV type (distance, angle, dihedral,
+coordination number, RMSD, path CV, etc.). The specific geometry of
+the CV is encoded entirely in the derivatives ds_k/dR_i that PLUMED
+provides. No CV-specific logic is needed here.
+
+Most atoms are not involved in any CV, so their bias force is zero.
+If an atom appears in multiple CVs, contributions are summed.
+
+
+Recovering the physical forces
+------------------------------
+    F_physical(t) = F_total(t) - F_bias(t)
+
+
+Unit conventions
+----------------
+This module is unit-agnostic — it performs only multiplication and
+subtraction. In the typical CP2K + PLUMED workflow with
+UNITS LENGTH=A ENERGY=eV:
+
+    Forces (CP2K, after conversion) : eV/Angstrom
+    PLUMED cv_force                 : eV/Angstrom
+    PLUMED cv_derivs                : dimensionless
+    Bias force = cv_force * cv_derivs : eV/Angstrom
+
+
+Functions
+---------
+    compute_bias_forces  — F_bias from CV forces and derivatives
+    remove_bias          — F_physical = F_total - F_bias
+    correction_summary   — statistics on the bias correction
+"""
+
+from __future__ import annotations
+
+import numpy as np
+
+
+def compute_bias_forces(
+    f_cv: np.ndarray,
+    ds_dR_list: list[np.ndarray],
+    atom_indices_list: list[list[int]],
+    natoms_total: int,
+) -> np.ndarray:
+    """
+    Compute bias forces on all atoms from one or more collective variables.
+
+    Applies the chain rule:
+
+                        K
+        F_bias_i(t) = SUM   f_k(t)  *  (ds_k / dR_i)(t)
+                       k=1
+
+    Parameters
+    ----------
+    f_cv : ndarray, shape (nframes, n_cvs)
+        Scalar force on each CV at each timestep.
+        Column k is f_k(t) = -dV_bias/ds_k.
+        For a single CV, shape (nframes, 1).
+
+    ds_dR_list : list of ndarray
+        One array per CV. Entry k has shape (nframes, natoms_in_cv_k, 3).
+        These are the derivatives of CV k with respect to each of its
+        constituent atoms' positions.
+
[…]
