+++ scripts/ForceCorrector.py
@@ -1,0 +1,1486 @@
+#!/usr/bin/env python3
+"""
+forceCorr.py — Remove PLUMED bias forces from CP2K trajectory forces
+               and produce ML-ready extended XYZ training data.
+
+Supports one or multiple collective variables (CVs). PLUMED writes all
+CV data as columns in shared files:
+
+    cv_force:   #! FIELDS time d1 d2 ...
+    cv_derivs:  #! FIELDS time parameter d1 d2 ...
+
+Each CV contributes a bias force via the chain rule:
+
+    F_bias = sum_k  f_cv_k * (ds_k / dR_k)
+
+The corrected (unbiased) forces are:
+
+    F_unbiased = F_total - F_bias
+
+When a CP2K position file (--pos) is provided, the script also reads
+coordinates and energies, converts energy from Hartree to eV, and
+writes a combined extended XYZ file ready for ML training.
+
+Usage examples
+--------------
+Produce extxyz for ML training (two CVs):
+    python forceCorr.py \\
+        --frc btd-frc-1.xyz \\
+        --pos btd-pos-1.xyz \\
+        --cv-force cv_force --cv-derivs cv_derivs \\
+        --cv-atoms 1,4 1,2 \\
+        --lattice 17.0 0 0  0 17.0 0  0 0 17.0 \\
+        --out-extxyz train.xyz
+
+With stress tensor and cell file:
+    python forceCorr.py \\
+        --frc btd-frc-1.xyz \\
+        --pos btd-pos-1.xyz \\
+        --cv-force cv_force --cv-derivs cv_derivs \\
+        --cv-atoms 1,4 1,2 \\
+        --cell-file btd-1.cell \\
+        --stress btd-stress-1.stress_tensor \\
+        --stress-unit au \\
+        --out-extxyz train.xyz
+
+Via YAML config file:
+    python forceCorr.py --config config.yaml
+
+Example config.yaml:
+    frc: btd-frc-1.xyz
+    pos: btd-pos-1.xyz
+    cv_force: cv_force
+    cv_derivs: cv_derivs
+    cvs:
+      - name: d1
+        atoms: [1, 4]
+      - name: d2
+        atoms: [1, 2]
+    lattice: [17.0, 0, 0, 0, 17.0, 0, 0, 0, 17.0]
+    # OR: cell_file: btd-1.cell
+    stress:
+      file: btd-stress-1.stress_tensor
+      unit: au       # au | eV/A^3 | GPa
+    output:
+      extxyz: train.xyz
+      xyz: forces_unbiased.xyz
+      flat: forces_unbiased.dat
+      bias_xyz: bias_forces.xyz
+    pbc: [true, true, true]
+    frames:
+      start: 0
+      end: null
+      stride: 1
+    time_tolerance: 1.0e-8
+"""
+
+from __future__ import annotations
+
+import argparse
+import logging
+import re
+import sys
+from dataclasses import dataclass, field
+from pathlib import Path
+from typing import Optional
+
+import numpy as np
+
+from ComputeBias import (
+    compute_bias_forces,
+    remove_bias,
+    correction_summary,
+)
+
+logger = logging.getLogger("forceCorr")
+
+
+# ──────────────────────────────────────────────────────────────────────
+# Unit conversions
+# ──────────────────────────────────────────────────────────────────────
+
+HARTREE_TO_EV = 27.211386245988
+BOHR_TO_ANG = 0.529177210903
+
+# Stress conversions TO eV/A^3
+_STRESS_CONVERSIONS = {
+    "ev/a^3": 1.0,
+    "ev/ang^3": 1.0,
[…]
