+++ docs/build/html/_sources/user_guide/index.rst.txt
@@ -1,0 +1,16 @@
+.. _user_guide:
+
+User Guide
+==========
+
+Concept-level documentation explaining how SPARC works and how to configure
+it for your system.
+
+.. toctree::
+   :maxdepth: 2
+
+   workflow_overview
+   active_learning
+   dft_engines
+   fine_tuning
+   plumed_enhanced_sampling

[…]
