+++ SPARC_v0.2_Updates.md
@@ -1,343 +1,0 @@
-# SPARC v0.2 — Updates and Current State
-
-**Smart Potential with Atomistic Rare Events & Continuous Learning**
-
----
-
-## Overview
-
-SPARC is an active learning workflow for developing reactive machine learning potentials.
-It orchestrates DFT calculations, ML model training, molecular dynamics simulations,
-and iterative refinement through Query-by-Committee (QbC) candidate selection.
-
----
-
-## Workflow Architecture
-
-```
-INPUT: YAML config → Load structures
-
-SECTION 1: Ab Initio MD (AIMD)
-├── DFT calculator setup (ORCA / VASP / CP2K / xTB)
-├── Initialize thermostat (Nose-Hoover / Langevin / NPT Berendsen / NVE)
-├── Temperature ramping (optional, linear VASP-style)
-├── PLUMED enhanced sampling (optional)
-└── Output: AIMD trajectory
-
-SECTION 2: MLIP Training
-├── Process AIMD trajectory → DeepMD npy format
-├── Fine-tune universal model (DPA-1/2/3, MACE) [optional]
-│   └── Or standard DeepMD from-scratch training
-├── Train ensemble (N models with different random seeds)
-└── Output: frozen_model_1.pth, frozen_model_2.pth, ...
-
-SECTION 3: ML-MD Simulation
-├── Load trained models via DeepPotential calculator
-├── Initialize thermostat + temperature ramping
-├── Run safety-checked MD (energy threshold, distance constraints)
-├── Checkpoint save/restore for restart
-├── Umbrella sampling (optional via PLUMED)
-└── Output: ML-MD trajectory
-
-SECTION 4: Active Learning Loop (iterative)
-├── Step 1: DFT Labeling — compute energies/forces for candidates
-├── Step 2: Retrain — combine all trajectories, retrain ensemble
-├── Step 3: ML-MD — run MD with updated models
-├── Step 4: Query-by-Committee — select new candidates
-└── Break when no candidates found or max iterations reached
-```
-
----
-
-## New Features & Changes
-
-### 1. Universal Model Fine-Tuning
-
-**New module:** `sparc/src/finetune.py`
-
-Fine-tune pre-trained universal ML potentials on system-specific DFT data instead of
-training from scratch. Supports two backends:
-
-| Backend | Command | Models | Output |
-|---------|---------|--------|--------|
-| DeePMD (DPA) | `dp --pt train --finetune model.pt` | DPA-1, DPA-2, DPA-3 | `.pth` |
-| MACE | `mace_run_train --foundation_model` | small, medium, large, custom `.model` | `.model` |
-
-**Key functions:**
-- `deepmd_finetune()` — Fine-tune DeePMD models with `--model-branch` support for multi-task models (e.g., DPA-3.2-5M with branches like `Omat24`, `Organic_Reactions`)
-- `mace_finetune()` — Fine-tune MACE foundation models with automatic data conversion (DeepMD npy → extxyz)
-- `finetune_training()` — Dispatcher that routes to the correct backend
-- `setup_MACE_calculator()` — ASE calculator setup for MACE models
-- `_convert_deepmd_to_extxyz()` — Data format converter with empty file detection
-
-**Configuration** (top-level in YAML):
-```yaml
-finetune:
-  enabled: True
-  model_type: "deepmd"              # "deepmd" or "mace"
-  pretrained_model: "DPA-3.2-5M.pt" # Model file or MACE name (small/medium/large)
-  model_branch: "Omat24"            # Multi-task branch (DPA-3 only)
[…]
