### 2026-05-02 #9468 @arfon
Hi @rahulumrao, sorry for the delay here. There are a few things that need addressing here:

**Typos/spelling:**

- workfow → workflow (Features and Implementation section)
- advance sampling technique → advanced sampling technique (Statement of need)

**Grammar:**

- "This allows user to use" → "This allows the user to use"
- "large-scale production run" → "production runs" (plural – I think?)

Finally, the Pietrucci & Andreoni (2011) bibliography entry contains a stray formatting artifact: Atomic structures<? Format?> And transformations. It looks like this should be cleaned up in the source BibTeX.
