+++ paper/paper.md
@@ -41,7 +41,7 @@ The motivation behind `SPARC` is to simplify the generation of high-quality trai
 
 Several groups have coupled AL with enhanced sampling or pathway exploration techniques. Vitartas et al. [@vitartas2025ArXiv] combined AL with well-tempered metadynamics to study organic reactions in both gas phase and solvent. Rivero et al. [@rivero2019JCP] trained PhysNet architecture together with an AL at 1000K MD for Diels–Alder and hydrogen transfer reactions. Ang et al. [@ang2021Chem] used SchNet combined with nudged elastic band calculations to explore solvent effects on pericyclic reactions. Parrinello and coworkers [@parrinello2020NatCom] coupled AL with variational enhanced sampling using DeePMD to investigate the phase diagram of gallium. Zhao et al. [@zhao2022umbrella] introduced a workflow combining umbrella sampling with AL to map out the solid-phase transition of GeSbTe. These studies highlight the promise of AL combined with enhanced sampling, but they remain highly tailored and driven by use case-specific implementations that can be difficult to generalize or reproduce.
 
-`SPARC` automates this process by integrating the PLUMED library [@tribello2014plumed] with AL protocol into a single modular workflow (see Figure 1). This allows user to use any advance sampling technique implemented in PLUMED to explore the configurational space and finding reactive configurations enabling scalable and reproducible MLIPs for generalized chemical environment. Other tools such as `janus-core`[@kasoar2024januscore], `ArcaNN`[@david2025arcann], and `FLARE`[@vandermause2020flare] also supports MLIP driven simulations with enhanced sampling. However, the integration of active learning with the PLUMED library into a single workflow remains an active area of development.
+`SPARC` automates this process by integrating the PLUMED library [@tribello2014plumed] with AL protocol into a single modular workflow (see Figure 1). This allows the user to use any advanced sampling technique implemented in PLUMED to explore the configurational space and finding reactive configurations enabling scalable and reproducible MLIPs for generalized chemical environment. Other tools such as `janus-core`[@kasoar2024januscore], `ArcaNN`[@david2025arcann], and `FLARE`[@vandermause2020flare] also supports MLIP driven simulations with enhanced sampling. However, the integration of active learning with the PLUMED library into a single workflow remains an active area of development.
 
 <!-- ![Schematic of the SPARC workflow showing the iterative loop of sampling, uncertainty estimation, and labeling.\label{fig:workflow}](workflow_diagram.svg) -->
 
@@ -52,7 +52,7 @@ Several groups have coupled AL with enhanced sampling or pathway exploration tec
 
 `SPARC` currently supports VASP [@hafner2008vasp] and CP2K [@hutter2014cp2k] for electronic structure calculation, although this could be readily expanded in the future. For MLIP training we use DeePMD-kit architecture [@wang2018deepmd], with the ensemble model approach for query-by-committee (QbC) uncertainty estimation [@miksch2021strategiesMLPs]. ML/MD simulations are run using ASE MD engine coupled together with both DeePMD and PLUMED calculator.
 
-ML/MD output is stored in ASE trajectory formats as this enables the broader ecosystem of analysis tools that already support ASE compatible formats. Since `SPARC` manages all stages via ASE, the workfow requires a Python environment with necessary dependencies. This makes the workflow highly portable across computing environments and suitable for both exploratory studies and large-scale production run.
+ML/MD output is stored in ASE trajectory formats as this enables the broader ecosystem of analysis tools that already support ASE compatible formats. Since `SPARC` manages all stages via ASE, the workflow requires a Python environment with necessary dependencies. This makes the workflow highly portable across computing environments and suitable for both exploratory studies and large-scale production runs.
 
[…]
