+++ paper/ref.bib
@@ -148,17 +148,17 @@
   doi={10.1021/acs.jctc.5b00846}
 }
 
-@article{pietrucci2011sprint,
-  title={Graph Theory Meets Ab Initio Molecular Dynamics: Atomic Structures<? format?> and Transformations at the Nanoscale},
-  author={Pietrucci, Fabio and Andreoni, Wanda},
-  journal={Physical review letters},
-  volume={107},
-  number={8},
-  pages={085504},
-  year={2011},
-  publisher={APS},
-  doi={10.1103/PhysRevLett.107.085504}
-}
+  @article{pietrucci2011sprint,
+    title={Graph Theory Meets Ab Initio Molecular Dynamics: Atomic Structures and Transformations at the Nanoscale},
+    author={Pietrucci, Fabio and Andreoni, Wanda},
+    journal={Physical review letters},
+    volume={107},
+    number={8},
+    pages={085504},
+    year={2011},
+    publisher={APS},
+    doi={10.1103/PhysRevLett.107.085504}
+  }
 
 @article{hafner2008vasp,
   title={Ab-initio simulations of materials using VASP: Density-functional theory and beyond},
