+++ README.md
@@ -7,7 +7,7 @@
 [![Ruff](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/astral-sh/ruff/main/assets/badge/v2.json)](https://github.com/astral-sh/ruff)
 [![Documentation Status](https://readthedocs.org/projects/sparc/badge/)](https://docs-sparc.readthedocs.io/en/latest/)
 [![CI](https://github.com/rahulumrao/sparc/actions/workflows/ci.yaml/badge.svg)](https://github.com/rahulumrao/sparc/actions)
-[![DOI](https://joss.theoj.org/papers/10.21105/joss.09468/status.svg)](https://doi.org/10.21105/joss.09468)
+[![DOI](https://joss.theoj.org/papers/10.21105/joss.09468/status.svg)](https://doi.org/10.21105/joss.09468) 
 
 **S**mart **P**otential with **A**tomistic **R**are Events and **C**ontinuous Learning
 
